% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(table = table)
}
\arguments{
\item{table}{insert the table (data frame).}
}
\value{
a processed data frame version of the original file.
}
\description{
Assigning the path's file_name +  processing columns & obtain a cleaned data frame
}
\examples{

\dontrun{

#Loading using a third party function and then cleaning the data frame.

path <- ".../earthquakes-2021-05-02_07-20-17_-0400.tsv"
table_original <- read.table(file = path, sep = '\t', header = TRUE)

#Insert table

eq_clean_data(table_original)}

}
