% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(map_table_example = rlang::.data, annot_col)
}
\arguments{
\item{map_table_example}{a data frame with Date, Longitude, Latitude, and Mag columns.}

\item{annot_col}{information to be displayed as popup when hovering.}
}
\value{
A map with earthquakes' circles.
}
\description{
It plots circles on a map to determine the eartquake location. The radius
of the circle is relative to its magnitude.
}
\examples{
#'
\dontrun{
# Creating a Dataset.
path <- "data/earthquakes-2021-05-02_07-20-17_-0400.tsv"
table_original <- read.table(file = path, sep = '\t', header = TRUE)
table <-eq_clean_data(table_original)
table_example <- dplyr::filter(table, lubridate::year(Date) > 2000,
                              Country \%in\% c("CHINA","JAPAN"))
table_example <- dplyr::select(table_example,Date,Country,Mag,Deaths,
Location.Name, Longitude, Latitude)

eq_map(table_example, annot_col = "Date")
}

}
